Remove LinearCoord in favor of Coord with a linearization step
**CRITICAL** bug correction - LinearCoord did not handled correctly unit
conversion or all processing functions which transform Linear to
NonLinear coord.

This PR remove LinearCoord object, but add a automatic
linearization/rounding step in Coord object.

- [x] Close the #643
- [x] Address a part of discussion #398
- [x] Tests have been added / modified
- [x] User-visible changes (including notable bug fixes) have been
documented in `docs/whatsnew/changelog.rst`.

+++ spectrochempy/core/readers/read_csv.py
@@ -109,7 +109,7 @@ def _read_csv(*args, **kwargs):
         delimiter = ";"
 
     d = [row for row in csv.reader(txt.splitlines(), delimiter=delimiter)]
-    d = np.array(d).T
+    d = np.array(d, dtype=float).T
 
     # First column is the x coordinates
     coordx = Coord(d[0])
